Refactor CoursesCategories component and pass categories as a prop
